CIRC-865 refactoring according to code review
